Basic Recursive Solutions Generator
Change p (I've kept this fixed at 2), q (I've used up to 257) and the rank (this usually works up to 16 or so before it takes too long) at the top of the first block. This will generate solutions and some information like the total number of solutions, the maximum denominator (as 2^kq^l), the greedy bound and whether the two agree.